+++ src/JsonApi/Request/JsonApiRelationshipInterface.php
@@ -1,0 +1,10 @@
+<?php
+namespace WoohooLabs\Yang\JsonApi\Request;
+
+interface JsonApiRelationshipInterface
+{
+    /**
+     * @return array
+     */
+    public function toArray();
+}