+++ src/JsonApi/Request/JsonApiToManyRelationship.php
@@ -1,0 +1,34 @@
+<?php
+namespace WoohooLabs\Yang\JsonApi\Request;
+
+class JsonApiToManyRelationship implements JsonApiRelationshipInterface
+{
+    /**
+     * @var array
+     */
+    protected $resourceIdentifiers = [];
+
+    public static function create()
+    {
+        return new self();
+    }
+
+    /**
+     * @param string $type
+     * @param string $id
+     */
+    public function addResourceIdentifier($type, $id)
+    {
+        $this->resourceIdentifiers[] = ["type" => $type, "id" => $id];
+    }
+
+    /**
+     * @return array
+     */
+    public function toArray()
+    {
+        return [
+            "data" => $this->resourceIdentifiers
+        ];
+    }
+}